Big Breakthrough for refund inspections
Better unlocking method for opening password protected xlsx files in
servant. This one will skip the decrypting step if file is not locked.

New function to check if program is installed on the system. It is
needed when file has to be opened in order to formulas be recalculated.

Property method 'quarter' added to inspector to get curent quarted.
Used to get right line number in loc table.

Changes in Inspector __init__ method. Input files are not being loaded
right at initialization of Inspector object, but rather in methods which
needs to operate in xlsx file. Reason is that in methods which are used
to check refunds in each xlsx table, if the formulas are not calculated
yet, file has to be opened with spreadsheet processor, save and then
open again with openpyxl.loadworkbook(). If we load workbook into
variable and only then we open file to do calculation of formulas, data
in variable with loaded workbook would not change.

And finally added methods for opening spreadsheet processor, simulate
keystrokes for saving documents, causing to recalculate furmulas and
values can then be read by openpyxl.

Also Vanguard class was renamed to Assembler, XScout to Scout and names
were updated in other files pointing to these classes.