Investigation for faulty values in refunds
First check which months have wrong value. This can be achieved
comparing sum of refunds in UP table with sum of refunds in LOC table.

Resaving of spreadsheet files is not done at initializing and object
Inspector, if it is needed. Then the workbook will be loaded to property
and can be accessed withount loading more that once.

insert_row() method in Servant class was copying values that are not
formulas, (payouts and fare for example). Fixed now.