Buttons run enforcer and inspect
Created simple template home page with buttons connected to running
script `enforcer.py` and `inspector.py`. Inspector needs to pass some
output message to page instead of standard IO in terminal. Now it prints
pandas dataframe, this has to be converted to some form of data that
could be shown on page. Same applies for enforcer.