Input from form send as string
CSV files are not needed to be uploaded to server, instead function in
views.py reads the content and sends it to enforcer as a long string.
This will be then processed to pandas dataframe as it done was before,
but from a file accessed directly by enforcer.

Static files folder is set up and loading is working as intended.